MutliPurposeScanProject/requirements.txt
pandas
scipy
pillow
numpy
scikit-learn 
pydicom
cryptography
opencv-python
matplotlib